added docs for video component and clean up
