+++ src/lib/Video/Controls.svelte
@@ -1,0 +1,67 @@
+<script>
+  import Fa from 'svelte-fa';
+  import { faReply, faPlay, faPause } from '@fortawesome/free-solid-svg-icons';
+  import { createEventDispatcher } from 'svelte';
+
+  const dispatch = createEventDispatcher();
+
+  export let paused;
+  export let controlsOpacity;
+  export let controlsPosition;
+  export let widthVideoContainer;
+  export let heightVideoContainer;
+  export let controlsBorderOffset;
+  export let resetCondition;
+  export let separateReplyIcon;
+  export let controlsColour;
+
+  function forwardBtnClick() {
+    paused === true ? (paused = false) : (paused = true);
+    dispatch('pausePlayEvent', {
+      text: paused,
+    });
+  }
+</script>
+
+<button
+  on:click="{forwardBtnClick}"
+  style="
+    opacity: {controlsOpacity}; 
+    top: {controlsPosition === 'top left' || controlsPosition === 'top right'
+    ? `${10}px`
+    : `${heightVideoContainer - 3}px`};
+    left: {controlsPosition === 'top left' || controlsPosition === 'bottom left'
+    ? `${10}px`
+    : `${widthVideoContainer - controlsBorderOffset}px`};
+    "
+>
+  {#if resetCondition}
+    <i class="play-pause-icon replay">
+      {#if separateReplyIcon}
+        <Fa icon="{faReply}" size="2x" color="{controlsColour}" />
+      {:else}
+        <Fa icon="{faPlay}" size="2x" color="{controlsColour}" />
+      {/if}
+    </i>
+  {:else if paused === false}
+    <i class="play-pause-icon pause">
+      <Fa icon="{faPause}" size="2x" color="{controlsColour}" />
+    </i>
+  {:else if paused === true}
+    <i class="play-pause-icon play">
+      <Fa icon="{faPlay}" size="2x" color="{controlsColour}" />
+    </i>
+  {:else}
+    error
+  {/if}
+</button>
+
+<style lang="scss">
+  button {
+    z-index: 2;
+    position: absolute;
+    cursor: pointer;
+    background-color: transparent;
+    border: none;
+  }
+</style>
