--- src/lib/VideoGeneral/docs.svx
@@ -1,107 +1,0 @@
----
-title: VideoGeneral
-description: Video component with lots of options
-slug: video-general
----
-
-<script>
-  import VideoGeneral from './index.svelte';
-  import DemoContainer from '../_docs/DemoContainer/index.svelte';
-  import LoremIpsum from './loremIpsum.svelte';
-</script>
-
-<section>
-
-    ## {title}
-
-    {description}
-
-</section>
-
-
-```svelte
-<script>
-  import { VideoGeneral } from '@reuters-graphics/graphics-svelte-components';
-</script>
-
-<VideoGeneral 
-  ariaDescription="Description of your video for screen readers."
-  src={'path-to-video-from-statics-folder'} // e.g. 'videos/azovstal-evacuation.mp4'
-  size='normal' // normal, wide, wider, widest or fluid
-/>
-```
-
-<DemoContainer>
-  <VideoGeneral 
-    ariaDescription="Description of your video for screen readers."
-    src={'videos/azovstal-evacuation.mp4'}
-    size='normal'
-  />
-</DemoContainer>
-
-
-
-
-
-<section>
-
-#### When to play and whether to loop
-Copy the source url for the Datawrapper chart in the `src` prop. 
-You can get this from the published url on Reuters graphics.
-- Publish the chart on Datawrapper.
-- Go to the **Datawrapper charts** Teams channel, wait for the graph to finish publishing.
-- Inside **Embed code (for developers only)**, find and copy the url inside the `src` tag.
-
-
-Note: There is no need to update the url if you update the chart inside Datawrapper. 
-Any changes will be automatically reflected. 
-
-</section>
-
-
-<section>
-
-#### Controls (play / pause)
-Copy the source url for the Datawrapper chart in the `src` prop. 
-You can get this from the published url on Reuters graphics.
-- Publish the chart on Datawrapper.
-- Go to the **Datawrapper charts** Teams channel, wait for the graph to finish publishing.
-- Inside **Embed code (for developers only)**, find and copy the url inside the `src` tag.
-
-
-Note: There is no need to update the url if you update the chart inside Datawrapper. 
-Any changes will be automatically reflected. 
-
-</section>
-
-
-<section>
-
-#### Videos with sound
-Copy the source url for the Datawrapper chart in the `src` prop. 
-You can get this from the published url on Reuters graphics.
-- Publish the chart on Datawrapper.
-- Go to the **Datawrapper charts** Teams channel, wait for the graph to finish publishing.
-- Inside **Embed code (for developers only)**, find and copy the url inside the `src` tag.
-
-
-Note: There is no need to update the url if you update the chart inside Datawrapper. 
-Any changes will be automatically reflected. 
-
-</section>
-
-
-<section>
-
-#### Other options
-Copy the source url for the Datawrapper chart in the `src` prop. 
-You can get this from the published url on Reuters graphics.
-- Publish the chart on Datawrapper.
-- Go to the **Datawrapper charts** Teams channel, wait for the graph to finish publishing.
-- Inside **Embed code (for developers only)**, find and copy the url inside the `src` tag.
-
-
-Note: There is no need to update the url if you update the chart inside Datawrapper. 
-Any changes will be automatically reflected. 
-
-</section>
