+++ src/lib/index.js
@@ -16,4 +16,4 @@ export { default as Scroller } from './Scroller/index.svelte';
 export { default as LottieScroller } from './LottieScroller/index.svelte';
 export { default as BeforeAfter } from './BeforeAfter/index.svelte';
 export { default as DatawrapperChart } from './DatawrapperChart/index.svelte';
-export { default as VideoGeneral } from './VideoGeneral/index.svelte';
+export { default as Video } from './Video/index.svelte';
